broke down the main.py
