+++ api/app/testdata/hotmaps-cdd_curr_adapted.tif
