+++ api/app/testdata/no_projection.tif
