+++ api/wsgi.py
@@ -1,6 +1,6 @@
 """WSGI entrypoint for production run
 """
-from main import app
+from app.app import app
 
 if __name__ == "__main__":
     app.run()
